+++ CHANGELOG.md
@@ -44,7 +44,8 @@ Bugsnag Notifiers on other platforms.
   `Bugsnag.notify()` callback.
   [#458](https://github.com/bugsnag/bugsnag-cocoa/pull/458)
 
-* `Bugsnag.stopSession()` is now `Bugsnag.pauseSession()`.
+* `Bugsnag.stopSession()` is now `Bugsnag.pauseSession()`.  This renaming has 
+   also been applied to the `BugsnagNotifier` and `BugsnagSessionTracker` classes.
   [#464](https://github.com/bugsnag/bugsnag-cocoa/pull/464)
 
 ## Bug fixes